requirements.txt
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.2
optuna==4.5.0
pytz==2025.2
requests==2.32.3
alpaca-py==0.19.0
coinbase-advanced-py==1.8.2
